add file upload handler

+++ app/Http/Controllers/PostsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 
@@ -55,14 +56,32 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999',
         ]);
+
+        // handle file upload
+        if($request->hasFile('cover_image')){
+            // Get filename with ext
+            $filenameWithExt = str_replace(' ','_', $request->file('cover_image')->getClientOriginalName());
+            // Get just Filename
+            $filename = \pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // get just Ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
 
         // Create Post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body  = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect()->route('posts.index')->with('success', 'Post Created');
@@ -109,13 +128,29 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999',
         ]);
 
+        if($request->hasFile('cover_image')){
+            // Get filename with ext
+            $filenameWithExt = str_replace(' ','_', $request->file('cover_image')->getClientOriginalName());
+            // Get just Filename
+            $filename = \pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // get just Ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
         // Create Post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body  = $request->input('body');
+        if ($request->hasFile('cover_image')) {
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect()->route('posts.index')->with('success', 'Post Updated');
@@ -130,8 +165,18 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
-        $post->delete();
 
+        // check for correct user
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect()->route('posts.index')->with('warning', 'Unauthorized Page');
+        }
+
+        if($post->cover_image != 'noimage.jpg') {
+            // delete image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
+        $post->delete();
         return redirect()->route('posts.index')->with('success', 'Post Deleted');
     }
 }
